List of Bar Plots

Create a list of bar plots for each categorical variable in a data frame. The proportions for each categorical variable are grouped by a different fill color.

Usage

categorical_bars(
  data_frame,
  cat_vars,
  target_col,
  title_size = 30,
  axis_size = 35
)

Arguments

data_frame

A data frame or data frame extension (e.g. a tibble).

cat_vars

A vector containing the string name(s) of each categorical variable to plot on the x-axis.

target_col

Column of interest to plot each variable against (object).

title_size

Size of each plot's title (Default = 30).

axis_size

Size of each plot's axes (Default = 35).

Value

A list containing ggplot2 object(s) (bar plots).

Examples

if (FALSE) { # \dontrun{
  # Generate bar plots for categorical variables from the mtcars dataset
  bar_plots <- categorical_bars(
    data_frame = mtcars,
    cat_vars = c("cyl", "gear"),
    target_col = "am",
    title_size = 25,
    axis_size = 20
  )
  # Display the first bar plot
  print(bar_plots[["cyl"]])
} # }