Logistic Regression Binary Classification Pipeline

Trains and fits a logistic regression model and cross-validates for optimal hyperparameter values.

Usage

lr_pipeline(
  data,
  target_col,
  vfolds = 5,
  grid_size = 10,
  tuning_metric,
  output_path
)

Arguments

data

A data frame or data frame extension (e.g. a tibble).

target_col

A string specifying the variable of interest.

vfolds

A number specifying the amount of folds used in k-fold cross-validation (Default = 5).

grid_size

A number specifying penalty values to test during model tuning (Default = 10).

tuning_metric

A string specifying the metric used to select for the most optimal model (e.g. "recall").

output_path

String path location to save the model as an RDS object.

Value

An RDS file containing the workflow object

Examples

if (FALSE) { # \dontrun{
  lr_pipeline(mtcars, "am", vfolds = 5, grid_size = 10, tuning_metric = "recall",
  output_path = "lasso_tuned_wflow.RDS")
} # }