Grid of Plots

Facet a list of ggplot2 bar plot objects into a single grid layout using patchwork. The number of columns in the grid can be specified with num_cols.

Usage

plots_grid(bar_plots, num_cols = 3)

Arguments

bar_plots

A list of ggplot2 bar plot objects.

num_cols

The number of columns in the grid (Default = 3).

Value

A patchwork / ggplot2 object showing all input plots arranged.

Examples

if (FALSE) { # \dontrun{
  # Generate example bar plots
  bar_plots <- categorical_bars(
    data_frame = mtcars,
    cat_vars = c("cyl", "gear"),
    target_col = "am"
    )
    
  # Combine plots into a grid
  grid_plot <- plots_grid(bar_plots, num_cols = 2)
} # }